Move booking views into seperate file
